Image processing optimizations
