+++ database/fileSave.js
@@ -1,5 +1,4 @@
 const fs = require('fs')
-const path = require('path')
 
 // Function used to save the labels to a JSON file.
 function saveLabels(sortedLabels, imageUrl, fileName) {
@@ -24,18 +23,6 @@ function saveLabels(sortedLabels, imageUrl, fileName) {
         })
       }
   })
-}
-
-// Function used to save the processed img to the processedImages folder
-function saveProcessedImg(file, canvas){
-  const processedImagePath = 'processedImages/processed' + path.basename(file.path)
-  const out = fs.createWriteStream(processedImagePath)
-  const stream = canvas.createJPEGStream()
-  stream.pipe(out)
-  out.on('finish', () => {
-      console.log('Processed image saved.')
-  })
-  return processedImagePath
 }
 
 module.exports = {
