--- views/style.css
@@ -1,297 +1,0 @@
-html, body {
-    margin: 0;
-    padding: 0;
-    font-family: "Poppins", sans-serif;
-}
-
-p {
-    padding: 0 10px;
-    line-height: 1.8;
-}
-
-ul li {
-    padding-right: 10px;
-    line-height: 1.6;
-}
-
-h3 {
-    padding: 5px 20px;
-    margin: 0;
-}
-
-div#header {
-    position: relative;
-    background-color: #0072CE;
-}
-
-.centreimage{
-    display: block;
-  margin-left: auto;
-  margin-right: auto;
-  width: 30%;
-  height: auto;
-}
-
-
-div#footer {
-    background: #42444e;
-    color: #fff;
-    position: fixed;
-    left:0;
-    bottom:0;
-    width:100%;
-    text-align: center;
-}
-
-div#footer p {
-    padding: 5px 5px;
-}
-
-div#extra {
-    clear: both;
-    width: 100%;
-}
-
-div#imageupload{
-    clear:both;
-    width:100%;
-}
-
-div#content{
-    padding-top: 10%;
-}
-
-.container{
-    background-color: #fff;
-    width:90%;
-    margin: 3.12em auto;
-    padding: 3.12em 1.25em;
-    border-radius: 0.43em;
-    box-shadow: 0 1.25em 2.18em #4d4e50;
-}
-
-input[type="file"]{
-    display: none;
-
-}
-
-label{
-    display:block;
-    position:relative;
-    background-color: #0072CE;
-    color: #fff;
-    font-size: 1.12em;
-    font-weight:500px;
-    text-align:center;
-    width:18.75em;
-    padding:1.12em 0;
-    margin:auto;
-    border-radius:0.31em;
-    cursor:pointer;
-}
-body {
-    background-color: #e9e9e9;
-    margin: 0px;
-    width: 100%;
-    margin: 0;
-    padding: 0;
-}
-
-ul li {
-    padding-right: 10px;
-    line-height: 1.6;
-}
-
-* {
-    box-sizing: border-box;
-  }
-
-  .custom-file-input {
-    color: transparent;
-  }
-  .custom-file-input::-webkit-file-upload-button {
-    visibility: hidden;
-  }
-  .custom-file-input::before {
-    content: 'Select some files';
-    color: black;
-    display: inline-block;
-    background: -webkit-linear-gradient(top, #f9f9f9, #e3e3e3);
-    border: 1px solid #02253e;
-    border-radius: 4px;
-    padding: 5px 10px;
-    cursor: pointer;
-    text-shadow: 1px 1px #fff;
-    font-weight: 700;
-    font-size: 10pt;
-    margin-left: 46%;
-    
-  }
-  .custom-file-input:hover::before {
-    border-color: black;
-  }
-
-.col-1 {width: 8.33%;}
-.col-2 {width: 16.66%;}
-.col-3 {width: 25%;}
-.col-4 {width: 33.33%;}
-.col-5 {
-    width: 41.66%;
-    align-content: center;
-    margin: 10px;
-}
-.col-52{
-    width: 400px;
-    align-content: center;
-    margin: 10px;
-}
-.col-6 {width: 50%;}
-.col-7 {width: 58.33%;}
-.col-8 {width: 66.66%;}
-.col-9 {width: 75%;}
-.col-10 {width: 83.33%;}
-.col-11 {width: 91.66%;}
-.col-12 {width: 100%;}
-
-[class*="col-"] {
-    float: left;
-    padding: 15px;
-    border: 1px solid rgb(0, 0, 0);
-}
-
-.row::after {
-    content: "";
-    clear: both;
-    display: table;
-  }
-
-.firstFooter {
-    background: #02253e;
-    color: #ffffff;
-    padding: 20px 10px;
-    box-sizing: border-box;
-    -ms-flex: 1;
-    flex: 1;
-}
-
-.secondFooter {
-    background: #f1f1f1;
-    color: #000000;
-    padding: 20px 10px;
-}
-
-.OurCompany {
-    width: 25%;
-    float: left;
-}
-
-.links2 {
-  width: 75%;
-  float: left;
-}
-
-.button1{
-    border-style: groove;
-    border-width: 2px;
-    border-color: black;
-    width: 85px;
-    height: 30px;
-    border-radius: 3px;
-    text-align: center;
-    color: rgb(0, 0, 0);
-    background-color: rgb(255, 255, 255);
-    text-decoration: none;
-}
-
-.button1:hover{
-    color: rgb(0, 0, 0);
-    cursor: pointer;
-    text-decoration: none;
-}
-
-.column-1{
-    text-align: center;
-    width: fit-content;
-}
-.firstFooterContainer{
-    -ms-flex: 1;
-    flex: 1;
-    text-align: center;
-    height: 200px;
-}
-
-[class*="column-"] {
-    float: left;
-}
-
-.ourCompany{
-    font-weight: 700;
-    padding-bottom: 10px;
-    text-decoration: solid;
-    font-size: large;
-}
-.footer-text{
-   color: rgb(255, 255, 255);
-   font-weight: 100;
-   line-height: 2;
-   width: 16.6%;
-   margin: 10px;
-   border:none;
-}
-
-[class*="footer-text"] {
-    float: left;
-}
-
-a:visited{
-    color: inherit;
-    text-decoration: none;
- }
- .backgroundHere{
-    background-image: url('Evergreen-Header-min.png');
-    height: 534px;
-    width: relative;
- }
- .header{
-    text-align: center;
-    font-weight: bolder;
- }
-
- #submit-button{
-    display:block;
-    position:relative;
-    background-color: #0072CE;
-    color: #fff;
-    font-size: 1.12em;
-    font-weight:500px;
-    text-align:center;
-    width:18.75em;
-    padding:1.12em 0;
-    margin:auto;
-    border-radius:0.31em;
-    cursor:pointer;
-    margin-top: 2%;
- }
-
- #submit-button.disabled{
-    display:block;
-    position:relative;
-    background-color: #0072CE;
-    color: #fff;
-    font-size: 1.12em;
-    font-weight:500px;
-    text-align:center;
-    width:18.75em;
-    padding:1.12em 0;
-    margin:auto;
-    border-radius:0.31em;
-    cursor:pointer;
-    margin-top: 2%;
-    opacity: 0.5;
-    cursor: not-allowed;
- }
-/* Takes the image upload and checks invalid, if invalid it changes style of submit */
- input[name = "image"]:invalid+input[name = "submit"]{
-    visibility: hidden;
- }
-
